Remove unit tracing support
Tracing came with the Rotonda base, and provided a way to see all
dynamic units in the system.  Nameshed has a static set of units, and
while it will track ongoing events internally, this implementation is
not quite suitable for it.  Removing support for tracing will help
clean up the codebase and tailor it to Nameshed.