Clean up most of the Rotonda base (#11)
* Remove unit tracing support

Tracing came with the Rotonda base, and provided a way to see all
dynamic units in the system.  Nameshed has a static set of units, and
while it will track ongoing events internally, this implementation is
not quite suitable for it.  Removing support for tracing will help
clean up the codebase and tailor it to Nameshed.

* Remove tokio metrics support

Nameshed is not I/O bound, and so tracking metrics from Tokio is not
expected to provide any significant benefits.

* [manager] Hard-code the unit configuration

Nameshed has a fixed set of units, so there's no need for an external
config file to determine them.  This change simply hardcodes the basic
configuration, as an incremental step towards removing configuration
support (for units!) entirely.

* Remove configuration support from the Rotonda base

As explained in the parent commit, Nameshed doesn't need the Rotonda
base's configuration file support.  This commit strips it out entirely
and replaces the parts still in use with hardcoded values for now.

* Remove unused modules

* [manager] Hard-code links to all units and targets

* [manager,comms] Remove link reports

* [comms] Remove support for cloning gates

This is an initial step towards removing gates entirely.  It greatly
simplifies the implementation of 'Gate::process()', allowing it to be
inlined into every unit.

* [manager,comms] Remove support for dynamic reconfiguration

* Send updates to central command with a dedicated channel

This makes 'Gate' mostly redundant; the only thing left is handling
termination gracefully.

* [comms] Remove suspension, subscription, and links

Now that updates for the central command are being sent via a manual
MPSC channel, nobody's making use of gate subscriptions.  Gates are
still used to send application commands to units.

* [comms] Send commands and termination via a custom MPSC channel

This eliminates the last remaining uses of 'Gate'.  At this point,
'Gate::process()' isn't even being called.

* [comms] Remove 'Gate'

* Remove all references to Roto

* [common] Remove 'FrimMap' and 'status_reporter'

'FrimMap' and 'DashMap' would perform about the same number of atomic
operations, and there was no performance improvement found.  Meanwhile
'papaya' is probably a much better approach to a read-heavy map.

* [manager::Component] stop tracking names

* Combine 'ZoneLoaderUnit' and 'ZoneLoader'

* [manager] Eliminate 'Coordinator' and 'Waitpoint'

'Manager' and 'CentralCommand' will eventually be merged, and will be
the single point of contact for all components, so there's never any
danger of a component sending a message to a mid-init component.

* [comms] Remove 'DirectUpdate' and some unused types