[zone] Add module 'contents'
This provides a mid-level API similar to 'zonetree'.  It has some
important differences:

- It does not support proper queries.
- It uses 'domain::new'.
- It can hold old versions of a zone.
- It provides direct access to the SOA record.
- It offers a synchronous API.

The idea is to enhance it to store both unsigned and signed versions of
a zone simultaneously, by separating the regular and signing-related
records.  This will provide a compact, uniform way to store zone data
before and after signing.