Reimplement IXFR and AXFR
This reimplements IXFR and AXFR based zone loading, strongly based on
the existing implementation in the zone maintainer.  The new code uses
'domain::new' wherever possible, and produces zone updates in the form
of 'ZoneContents'.

The interface should be changed in the future to improve performance by
processing the zone changes in parallel and on dedicated threads (using
'tokio::task::spawn_blocking()').  This would be useful for coalescing
IXFR diffs, computing diffs when an AXFR is received, and comparing
zones if the user forces a reload.

It's hard to tell what the performance impact of converting between
the old-base and 'domain::new' is.  It wouldn't be too hard to inline
the XFR interpretation (which is especially easy for AXFRs), and that
would eliminate most of the conversions.