Implement zone reloading
If a user is concerned that Nameshed's cache of loaded zone data is
incorrect, reloading can be used to fetch a new version of the zone
without relying on previous loaded zone data or to verify that the
local copy of the latest version of the zone is correct.

If ZONEMD is supported in the future, reloading could simply check the
ZONEMD record against the local copy to verify correctness.

Most of the time, Nameshed will be performing refreshes instead, which
do trust the local copy and can perform more efficient incremental zone
transfers.