Replace 'Manager' / 'Component' with 'Center'
While a lot of files and lines change, this is conceptually a simple
refactoring.  'Component' has now become 'Center', while 'Manager' is
replaced with a few more variables in 'main.rs'.  The most valuable
function in 'Manager' -- spawning all units -- is kept around in the
same source file.

The objective of this change is to provide a 'Center' to all units,
and enable them to start using the 'State' type (to access config,
zones, etc).  'State' will soon support saving/loading, thus powering
state persistence for all of Cascade.