Now that we have a zone add CLI command, don't forcibly add a zone based on env vars.
The ZL_IN_ZONE var is kept for now however as it is needed to know which
zone ZL_XFR_IN and PS_XFR_OUT relate to as those can't be configured yet
for a newly added zone.

Also replaces some hard-coded example.com references with the value of
ZL_IN_ZONE (which defaults to example.com still).