Remove obsoleted env var handling, and fix XFR/NOTIFY configuration.
- Use zone names as HashMap keys because they will match both
"example.com" and "example.com." while string matching will not, and
because the only way to add zones now is from the CLI which sends a
validated zone name.
- If XFR/NOTIFY settings exist for the zone being added _for the same
source_, use them, but if the source is different then use default
settings.
- Make logging for primary and secondary configuration consistent.
- Change the default mode of XFR from None to IXFR with AXFR fallback,
as if a primary is added we probably want to serve XFR by default to
secondaries.